refactor: move adapters to internal
