+++ internal/adapters/primary/http/handlers/generic_handler.go
@@ -1,9 +1,9 @@
 package handlers
 
 import (
+	msg "github.com/lechitz/AionApi/internal/adapters/primary/http/handlers/messages"
 	"net/http"
 
-	msg "github.com/lechitz/AionApi/adapters/primary/http/handlers/messages"
 	"github.com/lechitz/AionApi/pkg/utils"
 	"go.uber.org/zap"
 )
