+++ internal/adapters/primary/http/server/router.go
@@ -3,7 +3,7 @@ package server
 import (
 	"fmt"
 	"github.com/go-chi/chi/v5"
-	"github.com/lechitz/AionApi/adapters/primary/http/handlers"
+	"github.com/lechitz/AionApi/internal/adapters/primary/http/handlers"
 	inputHttp "github.com/lechitz/AionApi/internal/core/ports/input/http"
 	outputHttp "github.com/lechitz/AionApi/internal/core/ports/output/security"
 	"github.com/lechitz/AionApi/internal/platform/middleware/auth"
