+++ internal/platform/bootstrap/bootstrap.go
@@ -1,9 +1,9 @@
 package bootstrap
 
 import (
-	tokenadapter "github.com/lechitz/AionApi/adapters/secondary/cache"
-	dbadapter "github.com/lechitz/AionApi/adapters/secondary/db"
-	securityadapter "github.com/lechitz/AionApi/adapters/secondary/security"
+	tokenadapter "github.com/lechitz/AionApi/internal/adapters/secondary/cache"
+	dbadapter "github.com/lechitz/AionApi/internal/adapters/secondary/db"
+	securityadapter "github.com/lechitz/AionApi/internal/adapters/secondary/security"
 	"github.com/lechitz/AionApi/internal/core/ports/input/http"
 	"github.com/lechitz/AionApi/internal/core/ports/output/security"
 	"github.com/lechitz/AionApi/internal/core/service"
